requirements.txt
# StitchAdmin 2.0 Requirements
# Erstellt von Hans Hahn - Alle Rechte vorbehalten
# Python 3.11+ (getestet mit 3.11, 3.12, 3.13)

# ==========================================
# WEB FRAMEWORK
# ==========================================
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.3

# ==========================================
# DATENBANK
# FIX: SQLAlchemy 2.0.36+ für Python 3.13 Kompatibilität
# ==========================================
SQLAlchemy>=2.0.36

# ==========================================
# FORMULARE & VALIDIERUNG
# ==========================================
WTForms==3.1.2
email-validator==2.1.1

# ==========================================
# KONFIGURATION
# ==========================================
python-dotenv==1.0.1

# ==========================================
# SERVER (Optional für Produktion)
# ==========================================
gunicorn==22.0.0

# ==========================================
# BILD-VERARBEITUNG
# ==========================================
Pillow>=10.4.0

# ==========================================
# EXCEL & DOKUMENTE
# ==========================================
openpyxl==3.1.2
pandas>=2.2.0
xlrd==2.0.1

# ==========================================
# PDF-VERARBEITUNG
# ==========================================
PyPDF2==3.0.1
pdfplumber==0.10.3

# ==========================================
# STICKEREI-SPEZIFISCH
# ==========================================
pyembroidery==1.5.1

# ==========================================
# UTILITIES
# ==========================================
python-dateutil>=2.8.2
